refactor(circuit): split operator factories out of circuit_utils [Clean Code: SRP]
Move three closure-factory builders to a new
`scqubits/core/operator_factories.py`:

- `grid_operator_func_factory`
- `hierarchical_diagonalization_func_factory`
- `operator_func_factory`

All three are public; keep public deprecation shims in `circuit_utils.py`
so existing downstream imports keep working. The only internal caller
(`circuit_routines.py`) now imports from the new module directly.
